fixed more 64 bits warning
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@1442 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/textops/textops.c
@@ -1183,7 +1183,7 @@ static int fixup_privacy(void** param, int param_no)
 	return E_UNSPEC;
     }
     
-    *param = (void *)val;
+    *param = (void *)(long)val;
     return 0;
 }
 
@@ -1302,6 +1302,6 @@ static int is_privacy_f(struct sip_msg *msg, char *_privacy, char *str2 )
     if (parse_privacy(msg) == -1)
 	return -1;
 
-    return get_privacy_values(msg) & (unsigned int)_privacy ? 1 : -1;
+    return get_privacy_values(msg) & ((unsigned int)(long)_privacy) ? 1 : -1;
 
 }
